requirements.txt
Flask>=3.0,<4.0
Flask-JWT-Extended>=4.6
Flask-Cors>=4.0
pymongo[srv]>=4.6
python-dotenv>=1.0.1
structlog>=24.1
Flask-Limiter>=3.8
marshmallow>=3.21
firebase-admin>=6.5
requests>=2.31
APScheduler>=3.10
pytz>=2024.1
razorpay>=1.4.2
mistralai>=0.1.0
cloudinary>=1.36.0